Don't allow field scope to overwrite keys in type scope
