+++ packages/graphile-build/src/makeNewBuild.js
@@ -22,6 +22,8 @@ import type SchemaBuilder, {
   DataForType,
 } from "./SchemaBuilder";
 
+import extend from "./extend";
+
 const isString = str => typeof str === "string";
 const isDev = ["test", "development"].indexOf(process.env.NODE_ENV) >= 0;
 const debug = debugFactory("graphile-build");
@@ -178,16 +180,7 @@ export default function makeNewBuild(builder: SchemaBuilder): Build {
     getTypeByName(typeName) {
       return allTypes[typeName];
     },
-    extend<Obj1: *, Obj2: *>(base: Obj1, extra: Obj2): Obj1 & Obj2 {
-      const keysA = Object.keys(base);
-      const keysB = Object.keys(extra);
-      for (const key of keysB) {
-        if (keysA.indexOf(key) >= 0) {
-          throw new Error(`Overwriting key '${key}' is not allowed!`);
-        }
-      }
-      return Object.assign({}, base, extra);
-    },
+    extend,
     newWithHooks<T: GraphQLNamedType | GraphQLSchema, ConfigType: *>(
       Type: Class<T>,
       spec: ConfigType,
@@ -396,12 +389,10 @@ export default function makeNewBuild(builder: SchemaBuilder): Build {
                     }
                     return data;
                   },
-                  scope: Object.assign(
-                    {},
-                    scope,
-                    {
+                  scope: extend(
+                    extend(scope, {
                       fieldName,
-                    },
+                    }),
                     fieldScope
                   ),
                 });
@@ -486,12 +477,10 @@ export default function makeNewBuild(builder: SchemaBuilder): Build {
                   );
                 }
                 let context = Object.assign({}, commonContext, {
-                  scope: Object.assign(
-                    {},
-                    scope,
-                    {
+                  scope: extend(
+                    extend(scope, {
                       fieldName,
-                    },
+                    }),
                     fieldScope
                   ),
                 });
